requirements.txt
pymongo==4.10.1
APScheduler==3.10.1
google-search-results==2.4.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
